pyproject.toml
[tool.poetry]
name = "fastapi_python"
version = "0.1.0"
description = ""
authors = ["IohanaMaria <iohana.ma31@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
sqlalchemy = "^2.0.30"
pydantic = "^2.7.2"
alembic = "^1.13.1"
aiomysql = "^0.2.0"
pydantic-settings = "^2.2.1"
fastapi-pagination = "^0.12.24"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
serv = "serve:main"
create-migrations = "scripts:create_migrations"
run-migrations = "scripts:run_migrations"